+++ src/app/pages/manage/settings/assets-setting/imported-assets/imported-assets.component.html
@@ -1,0 +1,1 @@
+<p>imported-assets works!</p>
