+++ src/app/pages/manage/settings/assets-setting/imported-assets/imported-assets.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-imported-assets',
+  templateUrl: './imported-assets.component.html',
+  styleUrls: ['./imported-assets.component.css']
+})
+export class ImportedAssetsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
